requirements.txt
python-dotenv==1.0.0
requests==2.31.0
dnspython==2.4.2
groq==0.4.2
openpyxl==3.1.2
google-genai>=1.0.0
